Crate alloy_hardforks

alloy-hardforks

This crate contains definitions for Ethereum hardforks.

Re-exports

pub use ethereum::holesky;

pub use ethereum::hoodi;

pub use ethereum::mainnet;

pub use ethereum::sepolia;

pub use arbitrum::mainnet as arbitrum_mainnet;

pub use arbitrum::sepolia as arbitrum_sepolia;

Modules

arbitrum

Arbitrum hardfork starting points

ethereum

Ethereum hardfork starting points

Macros

hardfork

Macro that defines different variants of a chain specific enum. See crate::Hardfork as an example.

Structs

EnrForkIdEntry

Re-exported EIP-2124 forkid types. Represents a forward-compatible ENR entry for including the forkid in a node record via EIP-868. Forward compatibility is achieved via EIP-8.

EthereumChainHardforks

A type allowing to configure activation ForkConditions for a given list of EthereumHardforks.

ForkFilter

Re-exported EIP-2124 forkid types. Filter that describes the state of blockchain and can be used to check incoming ForkIds for compatibility.

ForkHash

Re-exported EIP-2124 forkid types. CRC32 hash of all previous forks starting from genesis block.

ForkId

Re-exported EIP-2124 forkid types. A fork identifier as defined by EIP-2124. Serves as the chain compatibility identifier.

ForkTransition

Re-exported EIP-2124 forkid types. Represents a transition from one fork to another

Head

Re-exported EIP-2124 forkid types. Describes the current head block.

Enums

EthereumHardfork

The name of an Ethereum hardfork.

ForkCondition

The condition at which a fork is activated.

ForkFilterKey

Re-exported EIP-2124 forkid types. How to filter forks.

ValidationError

Re-exported EIP-2124 forkid types. Reason for rejecting provided ForkId.

Traits

EthereumHardforks

Helper methods for Ethereum forks.

Hardfork

Generic hardfork trait.