alloy_hardforks

Enum EthereumHardfork

pub enum EthereumHardfork {

    Frontier,
    Homestead,
    Dao,
    Tangerine,
    SpuriousDragon,
    Byzantium,
    Constantinople,
    Petersburg,
    Istanbul,
    MuirGlacier,
    Berlin,
    London,
    ArrowGlacier,
    GrayGlacier,
    Paris,
    Shanghai,
    Cancun,
    Prague,
    Osaka,

}

The name of an Ethereum hardfork.

Variants

Frontier

Frontier: https://blog.ethereum.org/2015/03/03/ethereum-launch-process.

Homestead

Homestead: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/homestead.md.

Dao

The DAO fork: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/dao-fork.md.

Tangerine

Tangerine: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/tangerine-whistle.md.

SpuriousDragon

Spurious Dragon: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/spurious-dragon.md.

Byzantium

Byzantium: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/byzantium.md.

Constantinople

Constantinople: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/constantinople.md.

Petersburg

Petersburg: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/petersburg.md.

Istanbul

Istanbul: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/istanbul.md.

MuirGlacier

Muir Glacier: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/muir-glacier.md.

Berlin

Berlin: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/berlin.md.

London

London: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/london.md.

ArrowGlacier

Arrow Glacier: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/arrow-glacier.md.

GrayGlacier

Gray Glacier: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/gray-glacier.md.

Paris

Paris: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/paris.md.

Shanghai

Shanghai: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/shanghai.md.

Cancun

Cancun: https://github.com/ethereum/execution-specs/blob/master/network-upgrades/mainnet-upgrades/cancun.md

Prague

Prague.

Osaka

Osaka: https://eips.ethereum.org/EIPS/eip-7607

Implementations

impl EthereumHardfork

pub const VARIANTS: &'static [Self]

All hardfork variants

pub const fn name(&self) -> &'static str

Returns variant as str.

impl EthereumHardfork

pub fn activation_block(&self, chain: Chain) -> Option<u64>

Retrieves the activation block for the specified hardfork on the given chain.

pub const fn mainnet_activation_block(&self) -> Option<u64>

Retrieves the activation block for the specified hardfork on the Ethereum mainnet.

pub const fn sepolia_activation_block(&self) -> Option<u64>

Retrieves the activation block for the specified hardfork on the Sepolia testnet.

pub const fn arbitrum_sepolia_activation_block(&self) -> Option<u64>

Retrieves the activation block for the specified hardfork on the Arbitrum Sepolia testnet.

pub const fn arbitrum_activation_block(&self) -> Option<u64>

Retrieves the activation block for the specified hardfork on the Arbitrum One mainnet.

pub fn activation_timestamp(&self, chain: Chain) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the given chain.

pub const fn mainnet_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Ethereum mainnet.

pub const fn sepolia_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Sepolia testnet.

pub const fn holesky_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Holesky testnet.

pub const fn hoodi_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Hoodi testnet.

pub const fn arbitrum_sepolia_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Arbitrum Sepolia testnet.

pub const fn arbitrum_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Arbitrum One mainnet.

pub const fn mainnet() -> [(Self, ForkCondition); 18]

Ethereum mainnet list of hardforks.

pub const fn sepolia() -> [(Self, ForkCondition); 16]

Ethereum sepolia list of hardforks.

pub const fn holesky() -> [(Self, ForkCondition); 16]

Ethereum holesky list of hardforks.

pub const fn hoodi() -> [(Self, ForkCondition); 16]

Ethereum Hoodi list of hardforks.

pub const fn from_mainnet_block_number(num: u64) -> Self

Convert an u64 into an EthereumHardfork.

pub fn from_chain_and_timestamp(chain: Chain, timestamp: u64) -> Option<Self>

Reverse lookup to find the hardfork given a chain ID and block timestamp. Returns the active hardfork at the given timestamp for the specified chain.